+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 CHANGELOG
 ---------
 
+## 4.3.9
+* **Change** Error codes changed to 101-110 from 1-10
+* **Change** Started using PAN, regId, token, or CVV for simulating errors and delay instead of merchant ref_id
+
 ## 4.3.8
 * **Feature** Stage mode to start in http mode